refactor(ImportStateStatus): Change format of ImportStateStatus Enum
